Conflicts:
	modpk_potential.f90
	testm.f90

+++ modpk_backgrnd.f90
@@ -14,23 +14,23 @@ CONTAINS
   SUBROUTINE backgrnd
     INTEGER*4 :: i,j, rescl_count
 
-    DOUBLE PRECISION :: phi_init_trial(size(phi_init))
-    DOUBLE PRECISION :: alpha_e,dalpha,V_end,dv,dh
-    DOUBLE PRECISION :: a_end_inst
-    DOUBLE PRECISION :: V_i, ph, alpha_pivot, bb(size(phi_init))
-    DOUBLE PRECISION :: Np_last
+    real(dp) :: phi_init_trial(size(phi_init))
+    real(dp) :: alpha_e,dalpha,V_end,dv,dh
+    real(dp) :: a_end_inst
+    real(dp) :: V_i, ph, alpha_pivot, bb(size(phi_init))
+    real(dp) :: Np_last
 
     CHARACTER(16) :: fmt = '(a25,es10.3)'
 
     !MULTIFIELD
     CHARACTER(19) :: array_fmt
     CHARACTER(len=5) :: ci
-    
+
     write(ci, '(I5)'), size(phi_init)
     ci = adjustl(ci)
     array_fmt = '(a25,'//trim(ci)//'es10.3)'
     !END MULTIFIELD
-   
+
     !     Set initial conditions
     !
     !     y(1)=phi              dydx(1)=dphi/dx
@@ -49,13 +49,13 @@ CONTAINS
        !Check whether phi_infl_end is in the correct direction from phi_init
        IF(.NOT.(slowroll_infl_end)) THEN
           IF (phidot_sign(1) .GT.0 .AND. phi_init(1) .gt. phi_infl_end(1)) THEN
-             PRINT*, 'MODPK: Initial phi is smaller than final phi.' 
+             PRINT*, 'MODPK: Initial phi is smaller than final phi.'
              PRINT*, 'MODPK: Please check your initial conditions'
              PRINT*, 'MODPK: QUITTING'
              STOP
           ENDIF
           IF (phidot_sign(1) .LT.0 .AND. phi_init(1) .lt. phi_infl_end(1)) THEN
-             PRINT*, 'MODPK: Initial phi is larger than final phi.' 
+             PRINT*, 'MODPK: Initial phi is larger than final phi.'
              PRINT*, 'MODPK: Please check your initial conditions'
              PRINT*, 'MODPK: QUITTING'
              STOP
@@ -71,7 +71,7 @@ CONTAINS
           phi_init_trial=phi_init
           CALL trial_background(phi_init_trial, alpha_e, V_end)
           !! no rescaling implemented for multifield, so the code will exit here for multifield
-          IF ((pk_bad/=0) .OR. phi_init_trial(1).EQ.phi_init(1)) EXIT            
+          IF ((pk_bad/=0) .OR. phi_init_trial(1).EQ.phi_init(1)) EXIT
           rescl_count=rescl_count+1
           IF (rescl_count .EQ. 50) THEN
              pk_bad=2
@@ -85,17 +85,17 @@ CONTAINS
           phi_init = phi_init_trial
        END DO
     !MULTIFIELD
-    else 
+    else
        CALL trial_background(phi_init, alpha_e, V_end)
        if (pk_bad .ne. 0) then
           print*, 'MODPK: pk_bad = ', pk_bad
           stop
        end if
-    end if    
+    end if
     !END MULTIFIELD
 
     IF(pk_bad==0) THEN
-       !Matching condition 
+       !Matching condition
        V_i=pot(phi_init)
 
        !MULTIFIELD
@@ -125,7 +125,7 @@ CONTAINS
              N_pivot = -71.1d0-log(V_end/(M_Pl**4))/4.d0-log(k_pivot*Mpc2Mpl/H_pivot)
           end do
           a_end = EXP(alpha_e)*a_init
-       ELSE          
+       ELSE
           alpha_pivot = alpha_e-N_pivot
           i=locate(lna(1:nactual_bg),alpha_pivot)
           j=MIN(MAX(i-(4-1)/2,1),nactual_bg+1-4)
@@ -135,13 +135,13 @@ CONTAINS
           a_init=k_pivot*Mpc2Mpl/H_pivot/EXP(alpha_pivot)
           a_end=EXP(alpha_e)*a_init
           a_end_inst=EXP(-71.1d0+LOG(V_i/V_end)/4.d0+LOG((M_Pl**4)/V_i)/4.d0)
-          IF (a_end .GT. a_end_inst) THEN 
+          IF (a_end .GT. a_end_inst) THEN
              PRINT*,'MODPK: inflation ends too late with this N_pivot.'
              pk_bad=3
              RETURN
           ENDIF
        END IF
-              
+
        a_pivot = EXP(alpha_pivot)*a_init
        N_tot = alpha_e - lna(1)
 
@@ -153,7 +153,7 @@ CONTAINS
           WRITE(*, fmt) ' N_pivot =', N_pivot
           WRITE(*, fmt) ' N_tot =', N_tot
        end if
-             
+
        DO i=1,nactual_bg
           aharr(i)=LOG(a_init*EXP(lna(i))*hubarr(i))
        END DO
@@ -167,24 +167,24 @@ CONTAINS
     INTEGER*4, PARAMETER :: BNVAR=2
 
     !! MULTIFIELD
-    DOUBLE PRECISION, DIMENSION(:), INTENT(INOUT) :: phi_init_trial
-    DOUBLE PRECISION, DIMENSION(:) :: y(BNVAR*size(phi_init_trial)) 
+    real(dp), DIMENSION(:), INTENT(INOUT) :: phi_init_trial
+    real(dp), DIMENSION(:) :: y(BNVAR*size(phi_init_trial))
     !! END MUTLTIFIELD
 
-    DOUBLE PRECISION :: accuracy, h1, hmin, x1, x2
-    DOUBLE PRECISION :: alpha_e, dalpha, V_end, dv, ep 
-    DOUBLE PRECISION :: ph, alpha_pivot, aa(size(phi_init_trial)), bb(size(phi_init_trial))
-    DOUBLE PRECISION :: vv(nsteps) !!epsarr(nsteps), 
+    real(dp) :: accuracy, h1, hmin, x1, x2
+    real(dp) :: alpha_e, dalpha, V_end, dv, ep
+    real(dp) :: ph, alpha_pivot, aa(size(phi_init_trial)), bb(size(phi_init_trial))
+    real(dp) :: vv(nsteps) !!epsarr(nsteps),
 
-    DOUBLE PRECISION, DIMENSION(:) :: Vp(num_inflaton)    
-    DOUBLE PRECISION :: Vz, dotphi, thetaN, grad_V
+    real(dp), DIMENSION(:) :: Vp(num_inflaton)
+    real(dp) :: Vz, dotphi, thetaN, grad_V
 
     CHARACTER(16) :: fmt = '(a25,es10.3)'
 
     !MULTIFIELD
     CHARACTER(19) :: array_fmt
     CHARACTER(len=5) :: ci
-    
+
     write(ci, '(I5)'), size(phi_init_trial)
     ci = adjustl(ci)
     array_fmt = '(a25, '//trim(ci)//'es10.3)'
@@ -197,7 +197,7 @@ CONTAINS
 
     x1=0.0 !starting value
     x2=Nefold_max !ending value
-    
+
     !MULTIFIELD
     y(1 : size(y)/2) = phi_init_trial  !phi(x1)
     y(size(y)/2+1 : (size(y))) = -dVdphi(phi_init_trial)/3./h_init/h_init !dphi/dalpha(x1) slowroll approx
@@ -207,7 +207,7 @@ CONTAINS
     ode_underflow = .FALSE.
     ode_ps_output = .FALSE.
     ode_infl_end = .TRUE.
-    save_steps = .TRUE. 
+    save_steps = .TRUE.
     pk_bad = 0
 
     !MULTIFIELD
@@ -228,17 +228,17 @@ CONTAINS
     dxsav=1.d-7
     accuracy=1.0d-6
     hmin=0.0 !minimum stepsize
-    CALL odeint(y,x1,x2,accuracy,h1,hmin,bderivs,rkqs_r) 
+    CALL odeint(y,x1,x2,accuracy,h1,hmin,bderivs,rkqs_r)
 
-    IF(.NOT. ode_underflow) THEN    
-       lna(1:kount)=xp(1:kount)       
+    IF(.NOT. ode_underflow) THEN
+       lna(1:kount)=xp(1:kount)
        phiarr(:,1:kount)=yp(1:size(y)/2, 1:kount)
        dphiarr(:,1:kount)=yp(size(y)/2+1:size(y),1:kount)
        !MULTIFIELD
        DO i=1,kount
           vv(i) = pot(phiarr(:,i))
           hubarr(i) = getH(phiarr(:, i),dphiarr(:,i))
-          epsarr(i) = getEps(phiarr(:,i),dphiarr(:,i)) 
+          epsarr(i) = getEps(phiarr(:,i),dphiarr(:,i))
           sigma_arr(i) = sqrt(dot_product((phiarr(:,i)-phi_init),(phiarr(:,i)-phi_init)))
 
           !! compute d\theta/dN_e
@@ -251,7 +251,7 @@ CONTAINS
        !END MULTIFIELD
        !
        !     Determine the parameters needed for converting k(Mpc^-1) to K
-       !      
+       !
        nactual_bg=kount
        IF(slowroll_infl_end) THEN
           ep=1.d0
@@ -262,12 +262,12 @@ CONTAINS
           !MULTIFIELD
           CALL array_polint(epsarr(j:j+4), phiarr(:, j:j+4), ep, phi_infl_end, bb)
        ELSE
-          if (size(phi_init) .eq. 1) then 
+          if (size(phi_init) .eq. 1) then
              ep = abs(phi_init(1) - phi_infl_end(1))
              i = locate(sigma_arr(1:kount), ep)
              j = MIN(MAX(i-(4-1)/2,1),nactual_bg+1-4)
-             CALL polint(sigma_arr(j:j+4), lna(j:j+4), ep, alpha_e, dalpha)             
-             V_end = pot(phi_infl_end)             
+             CALL polint(sigma_arr(j:j+4), lna(j:j+4), ep, alpha_e, dalpha)
+             V_end = pot(phi_infl_end)
           else
              ep = delsigma
              i = locate(sigma_arr(1:kount), ep)
@@ -282,16 +282,16 @@ CONTAINS
        if (modpkoutput) WRITE(*,array_fmt) ' phi_end =', phi_infl_end
 
        IF (instreheat) THEN
-          !Set a plausible pivot (ignoring logarithmic V_k and V_end terms) for 
-          !the smallest k likely to be requested by CAMB. 
+          !Set a plausible pivot (ignoring logarithmic V_k and V_end terms) for
+          !the smallest k likely to be requested by CAMB.
           N_pivot=-LOG10(1.d-6)+60.
        ENDIF
-       
+
        !MULTIFIELD
        !
-       !Rescaling the initial value only implemented for single field cases. 
+       !Rescaling the initial value only implemented for single field cases.
        !For multi-field, one may need to specify whether to rescale a sigle compoent field,
-       !or to homogeneously rescale all the fields. It is better to work on a case by case basis. 
+       !or to homogeneously rescale all the fields. It is better to work on a case by case basis.
        !
 
        IF (size(phi_init) .eq. 1) THEN
@@ -299,11 +299,11 @@ CONTAINS
              IF ((potential_choice.eq.6).and.(vparams(1,1)<-2.d0)) THEN
                 phi_init_trial = phi_init*0.9d0
              ELSE
-                phi_init_trial = phi_init+(phi_init-phi_infl_end)*rescale_factor 
+                phi_init_trial = phi_init+(phi_init-phi_infl_end)*rescale_factor
              ENDIF
              RETURN
           END IF
-          
+
           IF (alpha_e.GT.N_pivot+55) THEN
              ep = alpha_e-(N_pivot+50)
              i = locate(lna(1:kount),ep)
@@ -316,17 +316,17 @@ CONTAINS
     ELSE
        pk_bad=4
     ENDIF
-    
-    !END MULTIFIELD 
+
+    !END MULTIFIELD
     RETURN
-  
+
   END SUBROUTINE trial_background
-  
+
   SUBROUTINE backgrnd_efold
-    
+
     INTEGER*4, PARAMETER :: BNVAR=2
-    DOUBLE PRECISION, DIMENSION(:) :: y(BNVAR*num_inflaton)
-    DOUBLE PRECISION :: accuracy, h1, hmin, x1, x2
+    real(dp), DIMENSION(:) :: y(BNVAR*num_inflaton)
+    real(dp) :: accuracy, h1, hmin, x1, x2
 
 
     h_init=SQRT(pot(phi_init)/(6.d0*M_Pl**2) * &
@@ -334,7 +334,7 @@ CONTAINS
 
     x1=0.0 !starting value
     x2=Nefold_max !ending value
-    
+
     y(1 : size(y)/2) = phi_init  !phi(x1)
     y(size(y)/2+1 : (size(y))) = -dVdphi(phi_init)/3./h_init/h_init !dphi/dalpha(x1) slowroll approx
 
@@ -342,7 +342,7 @@ CONTAINS
     ode_underflow = .FALSE.
     ode_ps_output = .FALSE.
     ode_infl_end = .TRUE.
-    save_steps = .TRUE. 
+    save_steps = .TRUE.
     pk_bad = 0
 
     IF(getEps(y(1:size(y)/2),y(size(y)/2+1:size(y))) .GT. 1.) THEN
@@ -361,17 +361,17 @@ CONTAINS
     dxsav=1.d-7
     accuracy=1.0d-6
     hmin=0.0 !minimum stepsize
-    CALL odeint(y,x1,x2,accuracy,h1,hmin,bderivs,rkqs_r) 
+    CALL odeint(y,x1,x2,accuracy,h1,hmin,bderivs,rkqs_r)
 
-    IF(.NOT. ode_underflow) THEN    
-       lna(1:kount)=xp(1:kount)       
+    IF(.NOT. ode_underflow) THEN
+       lna(1:kount)=xp(1:kount)
        nactual_bg=kount
     ELSE
        pk_bad = 4
     END IF
-    
+
     N_tot = lna(nactual_bg) - lna(1)
-     
+
   END SUBROUTINE backgrnd_efold
